N-API mapnik.VectorTile (WIP) [skip ci]
