+++ src/mapnik_vector_tile.cpp
@@ -10,6 +10,7 @@
 #include <mapnik/svg/output/svg_renderer.hpp>
 #endif
 
+//#define MAPNIK_VECTOR_TILE_LIBRARY
 #include "mapnik_vector_tile.hpp"
 //#include "vector_tile_compression.hpp"
 //#include "vector_tile_composite.hpp"
@@ -17,7 +18,7 @@
 //#include "vector_tile_projection.hpp"
 //#include "vector_tile_datasource_pbf.hpp"
 //#include "vector_tile_geometry_decoder.hpp"
-//#include "vector_tile_load_tile.hpp"
+#include "vector_tile_load_tile.hpp"
 #include "object_to_container.hpp"
 
 // mapnik
@@ -39,8 +40,6 @@
 #include <mapnik/projection.hpp>
 #include <mapnik/request.hpp>
 #include <mapnik/scale_denominator.hpp>
-#include <mapnik/util/geometry_to_geojson.hpp>
-#include <mapnik/util/feature_to_geojson.hpp>
 #include <mapnik/version.hpp>
 #if defined(GRID_RENDERER)
 #include <mapnik/grid/grid.hpp>         // for hit_grid, grid
@@ -1330,540 +1329,11 @@ Napi::Value VectorTile::painted(Napi::CallbackInfo const& info)
 }
 
 
-struct geometry_type_name
-{
-    template <typename T>
-    std::string operator () (T const& geom) const
-    {
-        return mapnik::util::apply_visitor(*this, geom);
-    }
 
-    std::string operator() (mapnik::geometry::geometry_empty const& ) const
-    {
-        // LCOV_EXCL_START
-        return "Empty";
-        // LCOV_EXCL_STOP
-    }
-
-    template <typename T>
-    std::string operator () (mapnik::geometry::point<T> const&) const
-    {
-        return "Point";
-    }
-
-    template <typename T>
-    std::string operator () (mapnik::geometry::line_string<T> const&) const
-    {
-        return "LineString";
-    }
-
-    template <typename T>
-    std::string operator () (mapnik::geometry::polygon<T> const&) const
-    {
-        return "Polygon";
-    }
-
-    template <typename T>
-    std::string operator () (mapnik::geometry::multi_point<T> const&) const
-    {
-        return "MultiPoint";
-    }
-
-    template <typename T>
-    std::string operator () (mapnik::geometry::multi_line_string<T> const&) const
-    {
-        return "MultiLineString";
-    }
-
-    template <typename T>
-    std::string operator () (mapnik::geometry::multi_polygon<T> const&) const
-    {
-        return "MultiPolygon";
-    }
-
-    template <typename T>
-    std::string operator () (mapnik::geometry::geometry_collection<T> const&) const
-    {
-        // LCOV_EXCL_START
-        return "GeometryCollection";
-        // LCOV_EXCL_STOP
-    }
-};
-
-template <typename T>
-static inline std::string geometry_type_as_string(T const& geom)
-{
-    return geometry_type_name()(geom);
-}
 /*
-struct geometry_array_visitor
-{
-    Napi::Array operator() (mapnik::geometry::geometry_empty const &)
-    {
-        // Removed as it should be a bug if a vector tile has reached this point
-        // therefore no known tests reach this point
-        // LCOV_EXCL_START
-        Napi::EscapableHandleScope scope(env);
-        return scope.Escape(Napi::Array::New(env));
-        // LCOV_EXCL_STOP
-    }
 
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::point<T> const & geom)
-    {
-        Napi::EscapableHandleScope scope(env);
-        Napi::Array arr = Napi::Array::New(env, 2);
-        (arr).Set(0, Napi::Number::New(env, geom.x));
-        (arr).Set(1, Napi::Number::New(env, geom.y));
-        return scope.Escape(arr);
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::line_string<T> const & geom)
-    {
-        Napi::EscapableHandleScope scope(env);
-        if (geom.empty())
-        {
-            // Removed as it should be a bug if a vector tile has reached this point
-            // therefore no known tests reach this point
-            // LCOV_EXCL_START
-            return scope.Escape(Napi::Array::New(env));
-            // LCOV_EXCL_STOP
-        }
-        Napi::Array arr = Napi::Array::New(env, geom.size());
-        std::uint32_t c = 0;
-        for (auto const & pt : geom)
-        {
-            (arr).Set(c++, (*this)(pt));
-        }
-        return scope.Escape(arr);
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::linear_ring<T> const & geom)
-    {
-        Napi::EscapableHandleScope scope(env);
-        if (geom.empty())
-        {
-            // Removed as it should be a bug if a vector tile has reached this point
-            // therefore no known tests reach this point
-            // LCOV_EXCL_START
-            return scope.Escape(Napi::Array::New(env));
-            // LCOV_EXCL_STOP
-        }
-        Napi::Array arr = Napi::Array::New(env, geom.size());
-        std::uint32_t c = 0;
-        for (auto const & pt : geom)
-        {
-            (arr).Set(c++, (*this)(pt));
-        }
-        return scope.Escape(arr);
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::multi_point<T> const & geom)
-    {
-        Napi::EscapableHandleScope scope(env);
-        if (geom.empty())
-        {
-            // Removed as it should be a bug if a vector tile has reached this point
-            // therefore no known tests reach this point
-            // LCOV_EXCL_START
-            return scope.Escape(Napi::Array::New(env));
-            // LCOV_EXCL_STOP
-        }
-        Napi::Array arr = Napi::Array::New(env, geom.size());
-        std::uint32_t c = 0;
-        for (auto const & pt : geom)
-        {
-            (arr).Set(c++, (*this)(pt));
-        }
-        return scope.Escape(arr);
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::multi_line_string<T> const & geom)
-    {
-        Napi::EscapableHandleScope scope(env);
-        if (geom.empty())
-        {
-            // Removed as it should be a bug if a vector tile has reached this point
-            // therefore no known tests reach this point
-            // LCOV_EXCL_START
-            return scope.Escape(Napi::Array::New(env));
-            // LCOV_EXCL_STOP
-        }
-        Napi::Array arr = Napi::Array::New(env, geom.size());
-        std::uint32_t c = 0;
-        for (auto const & pt : geom)
-        {
-            (arr).Set(c++, (*this)(pt));
-        }
-        return scope.Escape(arr);
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::polygon<T> const & poly)
-    {
-        Napi::EscapableHandleScope scope(env);
-        Napi::Array arr = Napi::Array::New(env, poly.size());
-        std::uint32_t index = 0;
-
-        for (auto const & ring : poly)
-        {
-            (arr).Set(index++, (*this)(ring));
-        }
-        return scope.Escape(arr);
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::multi_polygon<T> const & geom)
-    {
-        Napi::EscapableHandleScope scope(env);
-        if (geom.empty())
-        {
-            // Removed as it should be a bug if a vector tile has reached this point
-            // therefore no known tests reach this point
-            // LCOV_EXCL_START
-            return scope.Escape(Napi::Array::New(env));
-            // LCOV_EXCL_STOP
-        }
-        Napi::Array arr = Napi::Array::New(env, geom.size());
-        std::uint32_t c = 0;
-        for (auto const & pt : geom)
-        {
-            (arr).Set(c++, (*this)(pt));
-        }
-        return scope.Escape(arr);
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::geometry<T> const & geom)
-    {
-        // Removed as it should be a bug if a vector tile has reached this point
-        // therefore no known tests reach this point
-        // LCOV_EXCL_START
-        Napi::EscapableHandleScope scope(env);
-        return scope.Escape(mapnik::util::apply_visitor((*this), geom));
-        // LCOV_EXCL_STOP
-    }
-
-    template <typename T>
-    Napi::Array operator() (mapnik::geometry::geometry_collection<T> const & geom)
-    {
-        // Removed as it should be a bug if a vector tile has reached this point
-        // therefore no known tests reach this point
-        // LCOV_EXCL_START
-        Napi::EscapableHandleScope scope(env);
-        if (geom.empty())
-        {
-            return scope.Escape(Napi::Array::New(env));
-        }
-        Napi::Array arr = Napi::Array::New(env, geom.size());
-        std::uint32_t c = 0;
-        for (auto const & pt : geom)
-        {
-            (arr).Set(c++, (*this)(pt));
-        }
-        return scope.Escape(arr);
-        // LCOV_EXCL_STOP
-    }
-};
-
-template <typename T>
-Napi::Array geometry_to_array(mapnik::geometry::geometry<T> const & geom)
-{
-    Napi::EscapableHandleScope scope(env);
-    return scope.Escape(mapnik::util::apply_visitor(geometry_array_visitor(), geom));
-}
-
-struct json_value_visitor
-{
-    Napi::Object & att_obj_;
-    std::string const& name_;
-
-    json_value_visitor(Napi::Object & att_obj,
-                       std::string const& name)
-        : att_obj_(att_obj),
-          name_(name) {}
-
-    void operator() (std::string const& val)
-    {
-        (att_obj_).Set(Napi::New(env, name_), Napi::New(env, val));
-    }
-
-    void operator() (bool const& val)
-    {
-        (att_obj_).Set(Napi::New(env, name_), Napi::Boolean::New(env, val));
-    }
-
-    void operator() (int64_t const& val)
-    {
-        (att_obj_).Set(Napi::New(env, name_), Napi::Number::New(env, val));
-    }
-
-    void operator() (uint64_t const& val)
-    {
-        // LCOV_EXCL_START
-        (att_obj_).Set(Napi::New(env, name_), Napi::Number::New(env, val));
-        // LCOV_EXCL_STOP
-    }
-
-    void operator() (double const& val)
-    {
-        (att_obj_).Set(Napi::New(env, name_), Napi::Number::New(env, val));
-    }
-
-    void operator() (float const& val)
-    {
-        (att_obj_).Set(Napi::New(env, name_), Napi::Number::New(env, val));
-    }
-};
 */
-/**
- * Get a JSON representation of this tile
- *
- * @memberof VectorTile
- * @instance
- * @name toJSON
- * @param {Object} [options]
- * @param {boolean} [options.decode_geometry=false] return geometry as integers
- * relative to the tile grid
- * @returns {Object} json representation of this tile with name, extent,
- * version, and feature properties
- * @example
- * var vt = mapnik.VectorTile(10,131,242);
- * var buffer = fs.readFileSync('./path/to/data.mvt');
- * vt.setData(buffer);
- * var json = vectorTile.toJSON();
- * console.log(json);
- * // {
- * //   name: 'layer-name',
- * //   extent: 4096,
- * //   version: 2,
- * //   features: [ ... ] // array of objects
- * // }
- */
-Napi::Value VectorTile::toJSON(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-        //Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-    /*
-    bool decode_geometry = false;
-    if (info.Length() >= 1)
-    {
-        if (!info[0].IsObject())
-        {
-            Napi::Error::New(env, "The first argument must be an object").ThrowAsJavaScriptException();
-            return env.Undefined();
-        }
-        Napi::Object options = info[0].ToObject(Napi::GetCurrentContext());
 
-        if ((options).Has(Napi::String::New(env, "decode_geometry")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "decode_geometry"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::Error::New(env, "option 'decode_geometry' must be a boolean").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            decode_geometry = param_val.As<Napi::Boolean>().Value();
-        }
-    }
-
-    VectorTile* d = info.Holder().Unwrap<VectorTile>();
-    try
-    {
-        protozero::pbf_reader tile_msg = d->tile_->get_reader();
-        Napi::Array arr = Napi::Array::New(env, d->tile_->get_layers().size());
-        std::size_t l_idx = 0;
-        while (tile_msg.next(mapnik::vector_tile_impl::Tile_Encoding::LAYERS))
-        {
-            protozero::pbf_reader layer_msg = tile_msg.get_message();
-            Napi::Object layer_obj = Napi::Object::New(env);
-            std::vector<std::string> layer_keys;
-            mapnik::vector_tile_impl::layer_pbf_attr_type layer_values;
-            std::vector<protozero::pbf_reader> layer_features;
-            protozero::pbf_reader val_msg;
-            std::uint32_t version = 1;
-            while (layer_msg.next())
-            {
-                switch (layer_msg.tag())
-                {
-                    case mapnik::vector_tile_impl::Layer_Encoding::NAME:
-                        (layer_obj).Set(Napi::String::New(env, "name"), Napi::String::New(env, layer_msg.get_string()));
-                        break;
-                    case mapnik::vector_tile_impl::Layer_Encoding::FEATURES:
-                        layer_features.push_back(layer_msg.get_message());
-                        break;
-                    case mapnik::vector_tile_impl::Layer_Encoding::KEYS:
-                        layer_keys.push_back(layer_msg.get_string());
-                        break;
-                    case mapnik::vector_tile_impl::Layer_Encoding::VALUES:
-                        val_msg = layer_msg.get_message();
-                        while (val_msg.next())
-                        {
-                            switch(val_msg.tag())
-                            {
-                                case mapnik::vector_tile_impl::Value_Encoding::STRING:
-                                    layer_values.push_back(val_msg.get_string());
-                                    break;
-                                case mapnik::vector_tile_impl::Value_Encoding::FLOAT:
-                                    layer_values.push_back(val_msg.get_float());
-                                    break;
-                                case mapnik::vector_tile_impl::Value_Encoding::DOUBLE:
-                                    layer_values.push_back(val_msg.get_double());
-                                    break;
-                                case mapnik::vector_tile_impl::Value_Encoding::INT:
-                                    layer_values.push_back(val_msg.get_int64());
-                                    break;
-                                case mapnik::vector_tile_impl::Value_Encoding::UINT:
-                                    // LCOV_EXCL_START
-                                    layer_values.push_back(val_msg.get_uint64());
-                                    break;
-                                    // LCOV_EXCL_STOP
-                                case mapnik::vector_tile_impl::Value_Encoding::SINT:
-                                    // LCOV_EXCL_START
-                                    layer_values.push_back(val_msg.get_sint64());
-                                    break;
-                                    // LCOV_EXCL_STOP
-                                case mapnik::vector_tile_impl::Value_Encoding::BOOL:
-                                    layer_values.push_back(val_msg.get_bool());
-                                    break;
-                                default:
-                                    // LCOV_EXCL_START
-                                    val_msg.skip();
-                                    break;
-                                    // LCOV_EXCL_STOP
-                            }
-                        }
-                        break;
-                    case mapnik::vector_tile_impl::Layer_Encoding::EXTENT:
-                        (layer_obj).Set(Napi::String::New(env, "extent"), Napi::Number::New(env, layer_msg.get_uint32()));
-                        break;
-                    case mapnik::vector_tile_impl::Layer_Encoding::VERSION:
-                        version = layer_msg.get_uint32();
-                        (layer_obj).Set(Napi::String::New(env, "version"), Napi::Number::New(env, version));
-                        break;
-                    default:
-                        // LCOV_EXCL_START
-                        layer_msg.skip();
-                        break;
-                        // LCOV_EXCL_STOP
-                }
-            }
-            Napi::Array f_arr = Napi::Array::New(env, layer_features.size());
-            std::size_t f_idx = 0;
-            for (auto feature_msg : layer_features)
-            {
-                Napi::Object feature_obj = Napi::Object::New(env);
-                mapnik::vector_tile_impl::GeometryPBF::pbf_itr geom_itr;
-                mapnik::vector_tile_impl::GeometryPBF::pbf_itr tag_itr;
-                bool has_geom = false;
-                bool has_geom_type = false;
-                bool has_tags = false;
-                std::int32_t geom_type_enum = 0;
-                while (feature_msg.next())
-                {
-                    switch (feature_msg.tag())
-                    {
-                        case mapnik::vector_tile_impl::Feature_Encoding::ID:
-                            (feature_obj).Set(Napi::String::New(env, "id"),Napi::Number::New(env, feature_msg.get_uint64()));
-                            break;
-                        case mapnik::vector_tile_impl::Feature_Encoding::TAGS:
-                            tag_itr = feature_msg.get_packed_uint32();
-                            has_tags = true;
-                            break;
-                        case mapnik::vector_tile_impl::Feature_Encoding::TYPE:
-                            geom_type_enum = feature_msg.get_enum();
-                            has_geom_type = true;
-                            (feature_obj).Set(Napi::String::New(env, "type"),Napi::Number::New(env, geom_type_enum));
-                            break;
-                        case mapnik::vector_tile_impl::Feature_Encoding::GEOMETRY:
-                            geom_itr = feature_msg.get_packed_uint32();
-                            has_geom = true;
-                            break;
-                        case mapnik::vector_tile_impl::Feature_Encoding::RASTER:
-                        {
-                            auto im_buffer = feature_msg.get_view();
-                            (feature_obj).Set(Napi::String::New(env, "raster"),
-                                             Napi::Buffer::Copy(env, im_buffer.data(), im_buffer.size()));
-                            break;
-                        }
-                        default:
-                            // LCOV_EXCL_START
-                            feature_msg.skip();
-                            break;
-                            // LCOV_EXCL_STOP
-                    }
-                }
-                Napi::Object att_obj = Napi::Object::New(env);
-                if (has_tags)
-                {
-                    for (auto _i = tag_itr.begin(); _i != tag_itr.end();)
-                    {
-                        std::size_t key_name = *(_i++);
-                        if (_i == tag_itr.end())
-                        {
-                            break;
-                        }
-                        std::size_t key_value = *(_i++);
-                        if (key_name < layer_keys.size() &&
-                            key_value < layer_values.size())
-                        {
-                            std::string const& name = layer_keys.at(key_name);
-                            mapnik::vector_tile_impl::pbf_attr_value_type val = layer_values.at(key_value);
-                            json_value_visitor vv(att_obj, name);
-                            mapnik::util::apply_visitor(vv, val);
-                        }
-                    }
-                }
-                (feature_obj).Set(Napi::String::New(env, "properties"), att_obj);
-                if (has_geom && has_geom_type)
-                {
-                    if (decode_geometry)
-                    {
-                        // Decode the geometry first into an int64_t mapnik geometry
-                        mapnik::vector_tile_impl::GeometryPBF geoms(geom_itr);
-                        mapnik::geometry::geometry<std::int64_t> geom = mapnik::vector_tile_impl::decode_geometry<std::int64_t>(geoms, geom_type_enum, version, 0, 0, 1.0, 1.0);
-                        Napi::Array g_arr = geometry_to_array<std::int64_t>(geom);
-                        (feature_obj).Set(Napi::String::New(env, "geometry"), g_arr);
-                        std::string geom_type = geometry_type_as_string(geom);
-                        (feature_obj).Set(Napi::String::New(env, "geometry_type"), Napi::New(env, geom_type));
-                    }
-                    else
-                    {
-                        std::vector<std::uint32_t> geom_vec;
-                        for (auto _i = geom_itr.begin(); _i != geom_itr.end(); ++_i)
-                        {
-                            geom_vec.push_back(*_i);
-                        }
-                        Napi::Array g_arr = Napi::Array::New(env, geom_vec.size());
-                        for (std::size_t k = 0; k < geom_vec.size();++k)
-                        {
-                            (g_arr).Set(k, Napi::Number::New(env, geom_vec[k]));
-                        }
-                        (feature_obj).Set(Napi::String::New(env, "geometry"), g_arr);
-                    }
-                }
-                (f_arr).Set(f_idx++, feature_obj);
-            }
-            (layer_obj).Set(Napi::String::New(env, "features"), f_arr);
-            (arr).Set(l_idx++, layer_obj);
-        }
-        return arr;
-        return;
-    }
-    catch (std::exception const& ex)
-    {
-        // LCOV_EXCL_START
-        Napi::Error::New(env, ex.what()).ThrowAsJavaScriptException();
-        return env.Undefined();
-        // LCOV_EXCL_STOP
-    }
-    */
-}
 /*
 bool layer_to_geojson(protozero::pbf_reader const& layer,
                       std::string & result,
@@ -3267,256 +2737,7 @@ void VectorTile::EIO_AfterAddData(uv_work_t* req)
     delete closure;
 }
 */
-/**
- * Replace the data in this vector tile with new raw data (synchronous). This function validates
- * geometry according to the [Mapbox Vector Tile specification](https://github.com/mapbox/vector-tile-spec).
- *
- * @memberof VectorTile
- * @instance
- * @name setDataSync
- * @param {Buffer} buffer - raw data
- * @param {object} [options]
- * @param {boolean} [options.validate=false] - If true does validity checks mvt schema (not geometries)
- * Will throw if anything invalid or unexpected is encountered in the data
- * @param {boolean} [options.upgrade=false] - If true will upgrade v1 tiles to adhere to the v2 specification
- * @example
- * var data = fs.readFileSync('./path/to/data.mvt');
- * vectorTile.setDataSync(data);
- * // your custom code
- */
-Napi::Value VectorTile::setDataSync(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    Napi::EscapableHandleScope scope(env);
-    return env.Undefined();
-    /*
-    VectorTile* d = info.Holder().Unwrap<VectorTile>();
-    if (info.Length() < 1 || !info[0].IsObject())
-    {
-        Napi::TypeError::New(env, "first argument must be a buffer object").ThrowAsJavaScriptException();
 
-        return scope.Escape(env.Undefined());
-    }
-    Napi::Object obj = info[0].ToObject(Napi::GetCurrentContext());
-    if (!obj.IsBuffer())
-    {
-        Napi::TypeError::New(env, "first arg must be a buffer object").ThrowAsJavaScriptException();
-
-        return scope.Escape(env.Undefined());
-    }
-    std::size_t buffer_size = obj.As<Napi::Buffer<char>>().Length();
-    if (buffer_size <= 0)
-    {
-        Napi::Error::New(env, "cannot accept empty buffer as protobuf").ThrowAsJavaScriptException();
-
-        return scope.Escape(env.Undefined());
-    }
-    bool upgrade = false;
-    bool validate = false;
-    Napi::Object options = Napi::Object::New(env);
-    if (info.Length() > 1)
-    {
-        if (!info[1].IsObject())
-        {
-            Napi::TypeError::New(env, "second arg must be a options object").ThrowAsJavaScriptException();
-
-            return scope.Escape(env.Undefined());
-        }
-        options = info[1].ToObject(Napi::GetCurrentContext());
-        if ((options).Has(Napi::String::New(env, "validate")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "validate"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::TypeError::New(env, "option 'validate' must be a boolean").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            validate = param_val.As<Napi::Boolean>().Value();
-        }
-        if ((options).Has(Napi::String::New(env, "upgrade")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "upgrade"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::TypeError::New(env, "option 'upgrade' must be a boolean").ThrowAsJavaScriptException();
-
-                return scope.Escape(env.Undefined());
-            }
-            upgrade = param_val.As<Napi::Boolean>().Value();
-        }
-    }
-    try
-    {
-        d->clear();
-        merge_from_compressed_buffer(*d->get_tile(), obj.As<Napi::Buffer<char>>().Data(), buffer_size, validate, upgrade);
-    }
-    catch (std::exception const& ex)
-    {
-        Napi::Error::New(env, ex.what()).ThrowAsJavaScriptException();
-
-        return scope.Escape(env.Undefined());
-    }
-    return scope.Escape(env.Undefined());
-    */
-}
-/*
-typedef struct
-{
-    uv_work_t request;
-    VectorTile* d;
-    const char *data;
-    bool validate;
-    bool upgrade;
-    size_t dataLength;
-    bool error;
-    std::string error_name;
-    Napi::FunctionReference cb;
-    Napi::Persistent<v8::Object> buffer;
-} vector_tile_setdata_baton_t;
-*/
-
-/**
- * Replace the data in this vector tile with new raw data
- *
- * @memberof VectorTile
- * @instance
- * @name setData
- * @param {Buffer} buffer - raw data
- * @param {object} [options]
- * @param {boolean} [options.validate=false] - If true does validity checks mvt schema (not geometries)
- * Will throw if anything invalid or unexpected is encountered in the data
- * @param {boolean} [options.upgrade=false] - If true will upgrade v1 tiles to adhere to the v2 specification
- * @param {Function} callback
- * @example
- * var data = fs.readFileSync('./path/to/data.mvt');
- * vectorTile.setData(data, function(err) {
- *   if (err) throw err;
- *   // your custom code
- * });
- */
-Napi::Value VectorTile::setData(Napi::CallbackInfo const& info)
-{
-    Napi::Env env = info.Env();
-    return env.Undefined();
-
-    // ensure callback is a function
-/*
-    Napi::Value callback = info[info.Length() - 1];
-    if (!info[info.Length() - 1]->IsFunction())
-    {
-        return _setDataSync(info);
-        return;
-    }
-
-    if (info.Length() < 1 || !info[0].IsObject())
-    {
-        Napi::TypeError::New(env, "first argument must be a buffer object").ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-    Napi::Object obj = info[0].ToObject(Napi::GetCurrentContext());
-    if (!obj.IsBuffer())
-    {
-        Napi::TypeError::New(env, "first arg must be a buffer object").ThrowAsJavaScriptException();
-        return env.Undefined();
-    }
-
-    bool upgrade = false;
-    bool validate = false;
-    Napi::Object options = Napi::Object::New(env);
-    if (info.Length() > 1)
-    {
-        if (!info[1].IsObject())
-        {
-            Napi::TypeError::New(env, "second arg must be a options object").ThrowAsJavaScriptException();
-            return env.Undefined();
-        }
-        options = info[1].ToObject(Napi::GetCurrentContext());
-        if ((options).Has(Napi::String::New(env, "validate")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "validate"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::TypeError::New(env, "option 'validate' must be a boolean").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            validate = param_val.As<Napi::Boolean>().Value();
-        }
-        if ((options).Has(Napi::String::New(env, "upgrade")).FromMaybe(false))
-        {
-            Napi::Value param_val = (options).Get(Napi::String::New(env, "upgrade"));
-            if (!param_val->IsBoolean())
-            {
-                Napi::TypeError::New(env, "option 'upgrade' must be a boolean").ThrowAsJavaScriptException();
-                return env.Undefined();
-            }
-            upgrade = param_val.As<Napi::Boolean>().Value();
-        }
-    }
-
-    VectorTile* d = info.Holder().Unwrap<VectorTile>();
-
-    vector_tile_setdata_baton_t *closure = new vector_tile_setdata_baton_t();
-    closure->request.data = closure;
-    closure->validate = validate;
-    closure->upgrade = upgrade;
-    closure->d = d;
-    closure->error = false;
-    closure->cb.Reset(callback.As<Napi::Function>());
-    closure->buffer.Reset(obj.As<Napi::Object>());
-    closure->data = obj.As<Napi::Buffer<char>>().Data();
-    closure->dataLength = obj.As<Napi::Buffer<char>>().Length();
-    uv_queue_work(uv_default_loop(), &closure->request, EIO_SetData, (uv_after_work_cb)EIO_AfterSetData);
-    d->Ref();
-    return;
-*/
-}
-/*
-void VectorTile::EIO_SetData(uv_work_t* req)
-{
-    vector_tile_setdata_baton_t *closure = static_cast<vector_tile_setdata_baton_t *>(req->data);
-
-    if (closure->dataLength <= 0)
-    {
-        closure->error = true;
-        closure->error_name = "cannot accept empty buffer as protobuf";
-        return;
-    }
-
-    try
-    {
-        closure->d->clear();
-        merge_from_compressed_buffer(*closure->d->get_tile(), closure->data, closure->dataLength, closure->validate, closure->upgrade);
-    }
-    catch (std::exception const& ex)
-    {
-        closure->error = true;
-        closure->error_name = ex.what();
-    }
-}
-
-void VectorTile::EIO_AfterSetData(uv_work_t* req)
-{
-    Napi::HandleScope scope(env);
-    Napi::AsyncResource async_resource(__func__);
-    vector_tile_setdata_baton_t *closure = static_cast<vector_tile_setdata_baton_t *>(req->data);
-    if (closure->error)
-    {
-        Napi::Value argv[1] = { Napi::Error::New(env, closure->error_name.c_str()) };
-        async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 1, argv);
-    }
-    else
-    {
-        Napi::Value argv[1] = { env.Undefined() };
-        async_resource.runInAsyncScope(Napi::GetCurrentContext()->Global(), Napi::New(env, closure->cb), 1, argv);
-    }
-
-    closure->d->Unref();
-    closure->cb.Reset();
-    closure->buffer.Reset();
-    delete closure;
-}
-*/
 /**
  * Get the data in this vector tile as a buffer (synchronous)
  *
